Fix a use of `fresh-line'.
Needs to do it when the relevant stderr is in effect.

Fixes PR 13744, except that something like this is still bad:

  (with-output-to-file "/tmp/foo" #:exists 'append
    (lambda ()
      (parameterize ([current-error-port (current-output-port)])
        (/ 1 0))))

But I think that it's a bad idea for the error display handler to be
called with the locally set port -- described in PR 13974.

original commit: a2122418aee69b35e1dd08e57ef6f619c17680dd